requirements.txt
asgiref==3.2.5
autoflake==1.3.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-language-field==0.0.3
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.8
idna==2.9
isort==4.3.21
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
yapf==0.28.0
